feat: add migration to remove incompatible ldk payments
