+++ breez.go
@@ -37,7 +37,7 @@ func NewBreezService(mnemonic, apiKey, inviteCode, workDir string) (result lncli
 	}
 
 	//create dir if not exists
-	newpath := filepath.Join(".", workDir)
+	newpath := filepath.Join(workDir)
 	err = os.MkdirAll(newpath, os.ModePerm)
 	if err != nil {
 		return nil, err
